+++ app/src/main/java/com/appculinaryrecipes/FragmentLogin.java
@@ -1,0 +1,121 @@
+package com.appculinaryrecipes;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Toast;
+
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.databinding.DataBindingUtil;
+import androidx.fragment.app.Fragment;
+
+import com.appculinaryrecipes.databinding.LoginFragmentBinding;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GoogleAuthProvider;
+
+public class FragmentLogin extends Fragment {
+    private LoginFragmentBinding loginFragmentBinding;
+    private static final int RC_SIGN_IN = 100;
+    private GoogleSignInClient googleSignInClient;
+    private FirebaseAuth firebaseAuth;
+    private static final String TAG = "GOOGLE_SIGN_IN_TAG";
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        loginFragmentBinding = DataBindingUtil.inflate(inflater,R.layout.login_fragment,container,false);
+        loginFragmentBinding.setCallback(this);
+        View view = loginFragmentBinding.getRoot();
+        GoogleSignInOptions googleSignInOptions = new GoogleSignInOptions
+                .Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+        googleSignInClient = GoogleSignIn.getClient(getActivity().getApplicationContext(), googleSignInOptions);
+        firebaseAuth = FirebaseAuth.getInstance();
+        checkUser();
+
+        ActivityResultLauncher<Intent> someActivityResultLauncher = registerForActivityResult(
+                new ActivityResultContracts.StartActivityForResult(),
+                result -> {
+                    if (result.getResultCode() == Activity.RESULT_OK) {
+                        Intent data = result.getData();
+                        Task<GoogleSignInAccount> accountTask = GoogleSignIn.getSignedInAccountFromIntent(data);
+                        try {
+                            GoogleSignInAccount account = accountTask.getResult(ApiException.class);
+                            firebaseAuthWithGoogleAccount(account);
+                        } catch (Exception e) {
+                            System.out.println(e.getMessage());
+
+                        }
+                    }
+                });
+
+        loginFragmentBinding.signInButton.setOnClickListener(v -> {
+            Intent intent = googleSignInClient.getSignInIntent();
+            someActivityResultLauncher.launch(intent);
+        });
+        return view;
+    }
+
+    private void checkUser() {
+        FirebaseUser firebaseUser = firebaseAuth.getCurrentUser();
+        if (firebaseUser != null) {
+            startActivity(new Intent(getActivity(), MainActivity.class));
+        }
+    }
+
+    private void firebaseAuthWithGoogleAccount(GoogleSignInAccount account) {
+        AuthCredential credential = GoogleAuthProvider.getCredential(account.getIdToken(), null);
+        firebaseAuth.signInWithCredential(credential)
+                .addOnSuccessListener(authResult -> {
+                    startActivity(new Intent(getActivity(), MainActivity.class));
+                })
+                .addOnFailureListener(e -> {
+                    Toast.makeText(getActivity(), "Failure", Toast.LENGTH_LONG).show();
+                });
+    }
+
+    private void firebaseLogin(String email, String password){
+        firebaseAuth.signInWithEmailAndPassword(email, password)
+                .addOnCompleteListener(getActivity(), new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                            startActivity(new Intent(getActivity(), MainActivity.class));
+                        } else {
+                            Toast.makeText(getActivity(), "Failure", Toast.LENGTH_LONG).show();
+                        }
+                    }
+                });
+    }
+
+    public void onSubmit() {
+        if (!loginFragmentBinding.inputEmailText.getText().toString().isEmpty() &&
+                !loginFragmentBinding.editTextPassword.getText().toString().isEmpty()) {
+            firebaseLogin(loginFragmentBinding.inputEmailText.getText().toString(), loginFragmentBinding.editTextPassword.getText().toString());
+        }
+
+    }
+
+    public void onSignUpTextClick(){
+        //TODO: zaimplementować przełączanie się między widokami. Nie wiedziałem czy robimy na fragmentach czy przechodzimy przez Bundle normalnie
+
+    }
+}
